+++ tests/testthat/test-rd-s4.R
@@ -1,3 +1,14 @@
+test_that("invalid syntax generates useful warning", {
+  block <- "
+    #' A
+    #' @field
+    #' @slot
+    a <- function() {}
+  "
+
+  expect_snapshot(. <- roc_proc_text(rd_roclet(), block))
+})
+
 test_that("@fields creates a new section and lists fields", {
   out <- roc_proc_text(rd_roclet(), "
     #' Important class.